Update index.js
I realize that Amazon starts to block the scraper as early as the first 5 products (at least for me when I first started using this library). Changing the UA allows the scraper to circumvent the detection at least a little bit longer (works for more than 20+ products with 100 reviews each).